commander/requirements.txt
aiogram==3.4.1
docker==7.0.0
python-dotenv==1.0.1
requests==2.31.0